requirements.txt
tensorflow>=2.10.0
pillow>=9.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
imagehash>=4.3.0
requests>=2.28.0